Diagonal reject/accept optimization
For exponential, where the density function is convex everywhere, we can
short-circuit the exp() call half the time with a much cheaper test to
determine whether the point is above the diagonal: if it is, we reject
without needing to calculate exp(-x).

For normal, where the density function is concave until x=1 and convex
after that, we can do the same as above for [x_i, x_{i+1}] ranges above
1 (which covers 100/127 cases).  For x values below 1, (which covers
26/127 cases) where the density is concave, we can do a similar
short-circuiting by accepting (without calling exp) any points below the
diagonal.  The last case (1/127) is the one that includes the inflection
point: for that, we don't do any short-circuiting.  (The 128th case,
when we draw i = 0, involves a tail draw using the exponential
distribution).

Overall, this will reduce the number of exp calls by half for
exponential, and slightly less than half for normal, at the expense of
adding some (much cheaper) subtractions and multiplications in all the
cases where we might need to call exp().