Optimize diagonal calculation
By storing and using the U[0,1) we draw for y, we can avoid some
multiplications in the diagonal calculation which are just undoing the
U01 -> y transformation.

(Just to be extra safe, I drew 5 million normals and exponentials with
and without the change using mt19937 with a fixed seed, and got an
identical sequence of values).