Add tangency bounds to eliminate more exp calls
For exp and the convex portion of normal (i.e. where x is above 1) this
lets us accept draws without needing to check f(x); and for the concave
portion of normal, this lets us reject a good portion of the draws.

For exponential, we take a tangency to the exp(-x) curve at the
left-hand-side of the maybe-accept box; for normal, we take the tangency
to exp(-x^2/2) at the right-hand-side of the box (the right tangency is
considerably better for normal, and the left is marginally better for
exponential).

In terms of calls eliminated, this (combined with the 50% eliminated in
by the previous diagonal elimination) avoids exp calls 90.5% for normal,
and 98.67% of the calls for exponential.

Switching the normal elimination procedure to use two 2nd-order Taylor
series quadratics could increase the elimination rate for normal to
99.07%, but the added complexity of the code makes the average draw
speed slightly slower than this tangent+diagonal approach.