Use strict inequality for lower bound threshold
It's theoretically possible for floating point math to give us a case
where we are so close to the edge, or the linear approximation is so
good, that the lower bound check may accept a value that exp() would
not.  In practice, this seems an extremely unlikely event, and so this
isn't likely to make a difference, but using the same strict inequality
as the 'y < f(x)' on the next line does seem a little more
(mathematically) correct.