exp/a is just exp(a); document the tail procedure
exp/a is exactly the same as just generating an exp(a)--exponential
distribution just generates an exponential(1) and then divides it by a
when returning.  This commit makes that more obvious.

It also isn't immediately obvious how the tail generation is actually
working, so this also adds a comment detailing that this is really just
rejection sampling with a transformed condition.