Move mpl/bool include to detail/int_float_pair.hpp
normal_distribution doesn't actually use mpl::bool anymore--the code
that used it moved to detail/int_float_pair.hpp in
0b8f06ecfc6e5a1e488f121468e37c240471ead1; this include should have been
moved, too, but wasn't.

That said, this commit doesn't actually fix or change anything in the
current implementation: boost/mpl/bool.hpp is already included
indirectly from some of the other includes (e.g.
boost/random/uniform_int_distribution.hpp), but of course we shouldn't
depend on that, and this commit corrects that.